requirements.txt
opencv-python
torch
pillow
scikit-learn
numpy